move project to project.rs
