+++ reswmm-core/src/lib.rs
@@ -1,4 +1,5 @@
-extern crate enum_dispatch;
+/// Stormwater Management Model (SWMM) version 6
+
 pub extern crate furlong as units;
 pub mod xsection;
 pub mod element;
@@ -8,29 +9,7 @@ pub mod region;
 pub mod io;
 pub mod project;
 
-use std::{collections::HashMap};
-
-use element::UID;
-use link::{LinkElement, LinkKind};
-use node::NodeElement;
-
-pub struct Project {
-    nodes: HashMap<UID, NodeElement>,
-    links: HashMap<UID, LinkElement>
-}
-
-impl Project {
-    pub fn new() -> Self {
-        Project { nodes: HashMap::new(), links: HashMap::new() }
-    }
-    pub fn add_node(&mut self, node: NodeElement) -> Option<NodeElement> {
-        self.nodes.insert(node.uid, node)
-    }
-
-    pub fn add_link<L: Into<LinkKind>, S: ToString>(&mut self, uid: UID, name: S, link: L) -> Option<LinkElement> {
-        self.links.insert(uid, LinkElement::from((uid, name.to_string(), link.into())))
-    }
-}
+pub use project::Project;
 
 #[cfg(test)]
 mod test {
@@ -42,13 +21,17 @@ mod test {
         let mut prj = Project::new();
         let l = Conduit{length: 2.0};
         prj.add_link(1, "L-1", l);
-
-        #[cfg(feature="custom_links")]
-        {
-            let l2: link::CustomLink = link::CustomLink::new(Box::new(Conduit{length: 3.0}));
-            prj.add_link(2, "L-2", link::LinkKind::Custom(l2));
-        }
     }
+
+    #[cfg(feature="custom_links")]
+    #[test]
+    fn add_custom_link() {
+        let mut prj = Project::new();
+        let l2: link::CustomLink = link::CustomLink::new(Box::new(Conduit{length: 3.0}));
+        let old_link = prj.add_link(2, "L-2", link::LinkKind::Custom(l2));
+        assert!(old_link.is_none())
+    }
+
 }
 
 pub fn run() {
@@ -64,7 +47,7 @@ pub fn run() {
         prj.add_link(2, "L-2", link::LinkKind::Custom(l2));
     }
 
-    for link in prj.links {
+    for link in prj.links() {
         println!("{link:?}");
     }
 }
